pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spfluo"
dependencies = [
  "numpy",
  "scipy",
  "scikit-image",
  "pandas",
  "tqdm",
  "SimpleITK",
  "connected-components-3d",
  "albumentations",
  "psutil",
  "matplotlib",
  "array_api_compat",
]
version = "0.0.1"
authors = [
  { name="Jean Plumail", email="jplumail@unistra.fr" },
]
description = "SPFluo package"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
figtree = ["pyfigtree"]
gpu = [
  "cupy-wheel",
  "cucim",
  "torch",
  "torchvision",
  "tensorboard",
  "pynvml",
]
napari = [
  "napari[all] @ git+https://github.com/napari/napari.git@aa0330fcace8d5855fcd1e75c1fd4384b542bd4c",
]
dev = [
  "pytest",
]

[project.urls]
Repository = "https://github.com/jplumail/spfluo.git"

[tool.ruff]
select = [
    "E", "F", # default
    "I", # isort
]

[tool.ruff.isort]
known-first-party = ["spfluo"]
known-third-party = ["pyworflow", "pwfluo"]

[tool.ruff.per-file-ignores]
"spfluo/picking/*" = ["E501"]